refactor(project): re-export for simple importing with alias
