+++ src/interfaces/index.ts
@@ -1,0 +1,3 @@
+export type { default as ErrorMessage } from "./errorMessageProps";
+export type { default as PokemonProps } from "./pokemonProps";
+export type { default as Type } from "./type";
